+++ query-node/substrate-query-framework/cli/src/generate/utils.ts
