+++ query-node/substrate-query-framework/cli/test/fixtures/multiple-entities.graphql
@@ -1,17 +1,17 @@
-type Category {
+type Category @entity {
   id: ID!,
   name: String! @fulltext(query: "search"),
   parent_category: Category,
 }
 
-type Thread {
+type Thread @entity {
   id: ID!,
   category: Category!,
   title: String @fulltext(query: "search"),
   initial_body_text: String @fulltext(query: "search")
 }
 
-type Post {
+type Post @entity {
   id: ID!,
   thread: Thread!,
   index: Int!,
