+++ query-node/substrate-query-framework/cli/test/fixtures/multiple-queries.graphql
@@ -1,4 +1,4 @@
-type Membership {
+type Membership @entity {
   # Member's root account id
   accountId: String!
